Reorganized components folders. Fix URL for projects in service page.
