+++ components/service/Service.vue
@@ -62,8 +62,7 @@
             <nuxt-link
               :to="
                 localePath({
-                  name: 'projects-slug',
-                  params: { slug: work.url }
+                  name: `projects-${work.url}`
                 })
               "
             >
@@ -99,7 +98,7 @@
 </template>
 
 <script>
-import PageClaim from '~/components/PageClaim.vue'
+import PageClaim from '~/components/layout/PageClaim.vue'
 
 export default {
   components: {
